feat: 検索の見た目をコンポーネントとして外部化
